src/mkdoxy/requirements.txt
mkdocs
Jinja2
ruamel.yaml